refactor: co-locate page-specific components
Moved page-specific components from the central src/components/sections directory into a components sub-folder within each respective page directory in src/app. This improves modularity and makes the codebase easier to navigate and maintain.
