+++ src/app/about/components/education.tsx
